Lesson9/Task2/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
pip==20.2.3
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
setuptools==49.2.1
six==1.16.0
tornado==6.1
tzdata==2022.5
tzlocal==4.2
